refactor(worker-sessions): keep the new guards inside the file-count baselines
The package file-count baseline is exact and deletion-only, so a baselined
package cannot gain a file: both new test files were violations
(worker_sessions/internal/service 18 -> 19, orchestration/runtime 37 -> 38).
Two files had also crossed the 1000-line backend-size limit.

The terminal-reason guards move into control_production_boundary_test.go, which
despite its name is the observation-projection test file and is the surface
these guards assert. The dispatch-release guards move into
factory_move_work_test.go beside the other MoveWork tests. The two API
terminal-reason tests collapse into one table over (state, reason) pairs, which
also drops the duplicated request/response scaffolding.

observedTerminalCause moves from classify.go to observations.go: it is an
observation projection rather than a Workers-result classification, and it was
the reason classify.go crossed the limit.

No assertion is weakened and no coverage is dropped.
