+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/factory_move_work_test.go
@@ -2,15 +2,19 @@ package runtime
 
 import (
 	"context"
+	"errors"
+	"sync"
 	"testing"
+	"time"
 
 	"github.com/portpowered/infinite-you/internal/testutil/recordingfixtures"
-	"github.com/portpowered/infinite-you/pkg/services/work"
-
 	"github.com/portpowered/infinite-you/pkg/platform/logging"
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
+	factory "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
 	"github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/orchestrators/petri"
 	"github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/services/orchestration/state"
+	"github.com/portpowered/infinite-you/pkg/services/work"
+	"github.com/portpowered/infinite-you/pkg/services/workers"
 )
 
 func TestMoveWork_AcceptsWhileFactoryPaused(t *testing.T) {
@@ -135,4 +139,157 @@ func assertOperatorWorkStateChangeRecord(
 		}
 	}
 	t.Fatalf("records = %#v, want work-state change for %s %s -> %s", history.WorkStateChanges, workID, fromState, toState)
+}
+
+// dispatchReleaseObservationBudget bounds every wait in these guards. A wedged
+// dispatch is only released today by the two-hour workstation execution
+// timeout, so any bound this small distinguishes "the terminal outcome released
+// the dispatch" from "an outer timeout eventually did".
+const dispatchReleaseObservationBudget = 10 * time.Second
+
+// terminalOnReleaseExecutor holds one worker dispatch open until the test
+// releases it, then reports the exact terminal shape that a Worker Session
+// end-state hands back to Factory Runtime.
+//
+// The held phase reproduces the reported incident state: the dispatch is in
+// flight, it holds the work token, and `work move` is refused. The release
+// phase is the fix's contract -- a terminal worker outcome must retire the
+// dispatch on its own, without waiting on the execution timeout.
+type terminalOnReleaseExecutor struct {
+	started     chan struct{}
+	release     chan struct{}
+	startOnce   sync.Once
+	releaseGate sync.Once
+	cause       error
+}
+
+// releaseOnce ends the held dispatch and is safe to call more than once, so a
+// failed assertion can unwind the harness through the same path the happy case
+// uses.
+func (e *terminalOnReleaseExecutor) releaseOnce() {
+	e.releaseGate.Do(func() { close(e.release) })
+}
+
+func (e *terminalOnReleaseExecutor) Execute(
+	_ context.Context,
+	dispatch work.WorkDispatch,
+) (workers.WorkResult, error) {
+	e.startOnce.Do(func() { close(e.started) })
+	<-e.release
+	return workers.WorkResult{
+		DispatchID:   dispatch.DispatchID,
+		TransitionID: dispatch.TransitionID,
+		Outcome:      workers.OutcomeFailed,
+		Error:        e.cause.Error(),
+	}, e.cause
+}
+
+// TestMoveWork_CanceledWorkerSessionReleasesDispatch is the operator-cancelled
+// half of the wedged-token incident. `you worker-sessions cancel` drives the
+// attempt to ErrWorkstationDispatchCanceled; the work item's dispatch must
+// retire from that alone so `work move` succeeds immediately afterwards.
+func TestMoveWork_CanceledWorkerSessionReleasesDispatch(t *testing.T) {
+	assertTerminalWorkerOutcomeReleasesDispatch(
+		t,
+		"work-canceled-session-release",
+		workers.ErrWorkstationDispatchCanceled,
+	)
+}
+
+// TestMoveWork_ProcessGoneWorkerSessionReleasesDispatch is the process-gone
+// half. The operator kills the worker's OS process, the runner's bounded wait
+// turns that into a terminal failure, and the dispatch must retire without the
+// operator waiting out the two-hour execution timeout.
+func TestMoveWork_ProcessGoneWorkerSessionReleasesDispatch(t *testing.T) {
+	assertTerminalWorkerOutcomeReleasesDispatch(
+		t,
+		"work-process-gone-release",
+		workers.ErrWorkstationDispatchProcessGone,
+	)
+}
+
+// assertTerminalWorkerOutcomeReleasesDispatch proves the before/after of the
+// incident for one terminal cause: while the attempt is in flight `work move`
+// is refused with ErrMoveWorkInFlightDispatch, and once the attempt reaches
+// that terminal cause the dispatch releases and the same `work move` succeeds.
+func assertTerminalWorkerOutcomeReleasesDispatch(t *testing.T, workID string, cause error) {
+	t.Helper()
+
+	executor := &terminalOnReleaseExecutor{
+		started: make(chan struct{}),
+		release: make(chan struct{}),
+		cause:   cause,
+	}
+	harness := startServiceModeRunHarness(t,
+		withNet(buildSimpleNetWithFailureArc()),
+		withServiceMode(),
+		withWorkerExecutor("mock", executor),
+		withLogger(logging.NoopLogger{}),
+	)
+	// Registered before stop so it unwinds first: a failed assertion must never
+	// leave the executor blocked, or the harness shutdown hangs and reports a
+	// second, misleading failure.
+	t.Cleanup(harness.stop)
+	t.Cleanup(executor.releaseOnce)
+
+	ctx := context.Background()
+	if _, err := submitWorkRequests(ctx, harness.Factory, []work.SubmitRequest{{
+		WorkID:     workID,
+		WorkTypeID: "task",
+		TraceID:    "trace-" + workID,
+	}}); err != nil {
+		t.Fatalf("SubmitWorkRequest: %v", err)
+	}
+
+	select {
+	case <-executor.started:
+	case <-time.After(dispatchReleaseObservationBudget):
+		t.Fatalf("timed out waiting for the %s dispatch to reach the worker boundary", workID)
+	}
+	waitForAggregateSnapshot(t, harness.Factory, func(
+		snapshot *interfaces.EngineStateSnapshot[petri.MarkingSnapshot, *state.Net],
+	) bool {
+		return snapshot.InFlightCount > 0
+	})
+
+	// The wedged state the incident reported. A worker dispatch consumes the
+	// item's token into the transition for the life of the attempt, so the
+	// operator's recovery move is refused for as long as the attempt runs --
+	// which, before the runner's wait was bounded, meant until the two-hour
+	// execution timeout.
+	if _, err := harness.Factory.MoveWork(
+		ctx, workID, "done", work.WorkStateChangeSourceCLI, "",
+	); !errors.Is(err, factory.ErrMoveWorkNotFound) {
+		t.Fatalf("MoveWork while the attempt is in flight = %v, want %v", err, factory.ErrMoveWorkNotFound)
+	}
+
+	executor.releaseOnce()
+
+	// The terminal worker outcome alone must retire the dispatch. Nothing here
+	// waits on the workstation execution timeout.
+	waitForAggregateSnapshotWithTimeout(
+		t,
+		harness.Factory,
+		dispatchReleaseObservationBudget,
+		func(snapshot *interfaces.EngineStateSnapshot[petri.MarkingSnapshot, *state.Net]) bool {
+			return snapshot.InFlightCount == 0 &&
+				markingContainsWorkAtPlace(&snapshot.Marking, workID, "task:failed")
+		},
+	)
+
+	result, err := harness.Factory.MoveWork(ctx, workID, "done", work.WorkStateChangeSourceCLI, "")
+	if err != nil {
+		t.Fatalf("MoveWork after the attempt reached %v: %v", cause, err)
+	}
+	if result.FromState != "failed" || result.ToState != "done" {
+		t.Fatalf("move result = %#v, want failed -> done", result)
+	}
+
+	snapshot, err := harness.Factory.GetEngineStateSnapshot(ctx)
+	if err != nil {
+		t.Fatalf("GetEngineStateSnapshot: %v", err)
+	}
+	if !markingContainsWorkAtPlace(&snapshot.Marking, workID, "task:done") {
+		t.Fatalf("marking = %#v, want %s at task:done", snapshot.Marking.Tokens, workID)
+	}
 }
