+++ pkg/services/worker_sessions/transports/http/handler_test.go
@@ -54,73 +54,23 @@ func TestListWorkerSessionsBySessionIDProjectsPopulatedObservation(t *testing.T)
 	assertPopulatedListResponse(t, recorder.Body.Bytes(), total)
 }
 
-func TestListWorkerSessionsBySessionIDPreservesBoundedFailureKinds(t *testing.T) {
-	kinds := []workersessions.FailureCauseKind{
-		workersessions.FailureCauseRejected,
-		workersessions.FailureCauseIncompleteOutput,
-		workersessions.FailureCauseWorkersExecutionFailure,
-	}
-	observations := make([]workersessions.Observation, 0, len(kinds))
-	for _, kind := range kinds {
-		observations = append(observations, workersessions.Observation{
-			WorkerSessionID: "worker-session-" + string(kind),
-			AttemptID:       "attempt-" + string(kind),
-			WorkIDs:         []string{"work-1"},
-			State:           workersessions.StateFailed,
-			DurationBasis:   workersessions.DurationBasisRecordedTimestamps,
-			Transcript:      workersessions.TranscriptAvailabilityUnavailable,
-			Failure:         &workersessions.FailureCause{Kind: kind, Detail: "safe bounded detail"},
-		})
-	}
-	service := &fakeObservationService{result: workersessions.ListObservationsResult{Observations: observations}}
-	handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
-	recorder := httptest.NewRecorder()
-	request := httptest.NewRequest("GET", "/factory-sessions/session-1/worker-sessions?workId=work-1", nil)
-
-	handler.ListWorkerSessionsBySessionId(
-		recorder,
-		request,
-		factoryapi.SessionID("session-1"),
-		factoryapi.ListWorkerSessionsBySessionIdParams{WorkId: "work-1"},
-	)
-
-	if recorder.Code != http.StatusOK {
-		t.Fatalf("status = %d, want 200; body=%s", recorder.Code, recorder.Body.String())
-	}
-	var response factoryapi.ListWorkerSessionsResponse
-	if err := json.Unmarshal(recorder.Body.Bytes(), &response); err != nil {
-		t.Fatalf("decode response: %v", err)
-	}
-	if len(response.Sessions) != len(kinds) {
-		t.Fatalf("session count = %d, want %d", len(response.Sessions), len(kinds))
-	}
-	gotKinds := make(map[string]bool, len(response.Sessions))
-	for _, session := range response.Sessions {
-		if session.Failure == nil {
-			t.Fatalf("session %q has no failure classification", session.WorkerSessionId)
-		}
-		gotKinds[session.Failure.Kind] = true
-	}
-	for _, kind := range kinds {
-		if !gotKinds[string(kind)] {
-			t.Fatalf("response failure kinds = %#v, missing %q", gotKinds, kind)
-		}
-	}
-}
-
-// TestListWorkerSessionsBySessionIDNamesEachTerminalReason covers the API half
-// of the operator diagnosis path: an operator-canceled session must arrive
-// with its own named reason rather than a null failure, which reads exactly
-// like a cause that was never recorded.
-func TestListWorkerSessionsBySessionIDNamesEachTerminalReason(t *testing.T) {
+// TestListWorkerSessionsBySessionIDPreservesBoundedTerminalReasons covers the
+// API half of the operator diagnosis path. Every bounded reason must survive
+// the boundary paired with the state it ended in, including the operator
+// control outcomes: before those existed a canceled session arrived with a null
+// failure, which reads exactly like a cause that was never recorded.
+func TestListWorkerSessionsBySessionIDPreservesBoundedTerminalReasons(t *testing.T) {
 	reasons := []struct {
 		state workersessions.State
 		kind  workersessions.FailureCauseKind
 	}{
-		{state: workersessions.StateCanceled, kind: workersessions.FailureCauseOperatorCanceled},
-		{state: workersessions.StateTerminated, kind: workersessions.FailureCauseOperatorTerminated},
+		{state: workersessions.StateFailed, kind: workersessions.FailureCauseRejected},
+		{state: workersessions.StateFailed, kind: workersessions.FailureCauseIncompleteOutput},
+		{state: workersessions.StateFailed, kind: workersessions.FailureCauseWorkersExecutionFailure},
 		{state: workersessions.StateFailed, kind: workersessions.FailureCauseProcessGone},
 		{state: workersessions.StateFailed, kind: workersessions.FailureCauseTimeout},
+		{state: workersessions.StateCanceled, kind: workersessions.FailureCauseOperatorCanceled},
+		{state: workersessions.StateTerminated, kind: workersessions.FailureCauseOperatorTerminated},
 	}
 	observations := make([]workersessions.Observation, 0, len(reasons))
 	for _, reason := range reasons {
@@ -168,11 +118,9 @@ func TestListWorkerSessionsBySessionIDNamesEachTerminalReason(t *testing.T) {
 		if session.Failure == nil {
 			t.Fatalf("session ended by %q reports no named reason", reason.kind)
 		}
-		if session.Failure.Kind != string(reason.kind) {
-			t.Fatalf("session reason kind = %q, want %q", session.Failure.Kind, reason.kind)
-		}
-		if string(session.State) != string(reason.state) {
-			t.Fatalf("session state = %q, want %q", session.State, reason.state)
+		if session.Failure.Kind != string(reason.kind) || string(session.State) != string(reason.state) {
+			t.Fatalf("session reason = %q in state %q, want %q in state %q",
+				session.Failure.Kind, session.State, reason.kind, reason.state)
 		}
 	}
 }
